+++ app/Core/Models/Attachment.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace App\Core\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class Attachment extends Model
-{
-}