+++ database/factories/CategoryFactory.php
@@ -2,7 +2,7 @@
 
 use Faker\Generator as Faker;
 
-$factory->define(\App\Core\Models\Category::class, function (Faker $faker) {
+$factory->define(\App\Base\Models\Category::class, function (Faker $faker) {
     return [
         'name' => $faker->name,
     ];