+++ tests/Feature/HomePageTest.php
@@ -3,8 +3,8 @@
 namespace Tests\Feature;
 
 use Tests\TestCase;
-use App\Core\Models\Task;
-use App\Core\Models\Project;
+use App\TaskManager\Models\Task;
+use App\Base\Models\Project;
 use Laravel\Passport\Passport;
 use Illuminate\Support\Facades\Event;